publishNFT is broken up into testable units
