+++ src/components/HomeParts/FeaturedArt/index.tsx
@@ -3,10 +3,10 @@ import styled from 'styled-components'
 import { palette, breakpoint } from '@theme'
 import { Shimmer } from '@components'
 import { rgba, LayoutContext } from '@lib'
-import { IGrantsWithProjectFragment } from '@types'
+import { IGrantFragment } from '@types'
 
 type IFeaturedArt = {
-  grant?: IGrantsWithProjectFragment
+  grant?: IGrantFragment
   loading: boolean
 }
